INS-27: precprocessor for goplugin makes interface wrapper for classes
