+++ logicrunner/goplugin/goplugin_test.go
@@ -3,10 +3,6 @@ package goplugin
 import (
 	"testing"
 
-	"plugin"
-
-	"reflect"
-
 	"github.com/insolar/insolar/logicrunner"
 
 	"bytes"
@@ -43,19 +39,4 @@ func TestHelloWorld(t *testing.T) {
 	}
 	//	if ret == logicrunner.Arguments{} // IDK, lets decide what must be here
 	t.Log(ret)
-}
-
-var PATH = "./testplugins"
-
-func TestConfigLoad(t *testing.T) {
-	plugin, err := plugin.Open(PATH + "/secondary.so")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	hw, err := plugin.Lookup("EXP")
-	r := reflect.ValueOf(hw)
-	m := r.MethodByName("Hello")
-	ret := m.Call([]reflect.Value{})
-	t.Logf("%+v", ret)
 }
