--- logicrunner/goplugin/testplugins/secondary/src/awkwardcontract/main.go
@@ -1,52 +1,0 @@
-package awkwardcontract
-
-// command for build this file go build -buildmode=plugin main.go
-
-import "errors"
-
-// @inscontract
-type HelloWorlder struct { //Docccc
-	Greeted int
-}
-
-// @method
-func (hw *HelloWorlder) Hello() (string, error) {
-	hw.Greeted++
-	return "Hello world 2", nil
-}
-
-func (hw *HelloWorlder) Fail() (string, error) {
-	hw.Greeted++
-	return "", errors.New("We failed 2")
-}
-
-func (hw *HelloWorlder) Echo(s string) (string, error) {
-	hw.Greeted++
-	return s, nil
-}
-
-func (hw HelloWorlder) ConstEcho(s string) (string, error) {
-	return s, nil
-}
-
-func JustExportedStaticFunction(int, int) {}
-
-// generated
-
-func __INSMETHOD__Hello(_self *HelloWorlder) {
-	_self.Hello()
-}
-
-func __INSMETHOD__Fail(_self *HelloWorlder) {
-	_self.Fail()
-}
-
-func __INSMETHOD__Echo(_self *HelloWorlder, s interface{}) {
-	_self.Echo(s.(string))
-}
-
-func __INSMETHOD__ConstEcho(_self *HelloWorlder, s interface{}) {
-	_self.ConstEcho(s.(string))
-}
-
-var __INSEXPORTHelloWorlder HelloWorlder
